project_setup/settings.py
from pathlib import Path
from decouple import config, Csv
import os
import platform

# Build paths inside the project like this: BASE_DIR / 'subdir'.
BASE_DIR = Path(__file__).resolve().parent.parent


# Quick-start development settings - unsuitable for production
# See https://docs.djangoproject.com/en/4.2/howto/deployment/checklist/

# SECURITY WARNING: keep the secret key used in production secret!
SECRET_KEY = config('SECRET_KEY')

# SECURITY WARNING: don't run with debug turned on in production!
DEBUG = True

ALLOWED_HOSTS = [
    '*',
]

# View de Login
LOGIN_URL = 'signin'

# Application definition
DJANGO_APPS = [
    'django.contrib.admin',
    'django.contrib.auth',
    'django.contrib.contenttypes',
    'django.contrib.sessions',
    'django.contrib.messages',
    'django.contrib.staticfiles',
]

THIRD_PARTY_APPS = [
    'django_browser_reload',
    'tailwind',
    'app_tailwind',
]

MY_APPS = [
    'app_auth.apps.AppAuthConfig',                      # CONTROLE DE AUTENTICAÇÃO (LOGIN E CADASTRO)
    'app_dashboard.apps.AppDashboardConfig',            # ACESSO AO DASHBOARD
    'app_clients.apps.AppClientsConfig',                # CADASTRO E GERENCIAMENTO DE CLIENTS
    'app_stock.apps.AppStockConfig',                    # CONTROLE DE ESTOQUE (ENTRADAS E SAIDAS)
    'app_suppliers.apps.AppSuppliersConfig',            # CADASTRO E GERENCIAMENTO DE FORNECEDORES
    'app_products.apps.AppProductsConfig',              # CADASTRO E GERENCIAMENTO DE PRODUTOS
    'app_settings.apps.AppSettingsConfig',              # CONFIGURAÇÕES DO USUÁRIO
    'app_queries.apps.AppQueriesConfig',                # REALIZAÇÃO DE CONSULTAS
]

INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + MY_APPS

MIDDLEWARE = [
    'django.middleware.security.SecurityMiddleware',
    'django.contrib.sessions.middleware.SessionMiddleware',
    'django.middleware.common.CommonMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
    'django.contrib.auth.middleware.AuthenticationMiddleware',
    'django.contrib.messages.middleware.MessageMiddleware',
    'django.middleware.clickjacking.XFrameOptionsMiddleware',
    'django_browser_reload.middleware.BrowserReloadMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
]

ROOT_URLCONF = 'project_setup.urls'

TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [os.path.join(BASE_DIR, 'templates')],
        'APP_DIRS': True,
        'OPTIONS': {
            'context_processors': [
                'django.template.context_processors.debug',
                'django.template.context_processors.request',
                'django.contrib.auth.context_processors.auth',
                'django.contrib.messages.context_processors.messages',
                'django.template.context_processors.csrf',
            ],
        },
    },
]

WSGI_APPLICATION = 'project_setup.wsgi.application'


# Database
# https://docs.djangoproject.com/en/4.2/ref/settings/#databases

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.sqlite3',
        'NAME': BASE_DIR / 'db.sqlite3',
    }
}


# Password validation
# https://docs.djangoproject.com/en/4.2/ref/settings/#auth-password-validators

AUTH_PASSWORD_VALIDATORS = [
    {
        'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator',
    },
    {
        'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator',
    },
]


# Internationalization
# https://docs.djangoproject.com/en/4.2/topics/i18n/

LANGUAGE_CODE = 'pt-br'

TIME_ZONE = 'America/Sao_Paulo'

USE_I18N = True

USE_TZ = True


# Static files (CSS, JavaScript, Images)
# https://docs.djangoproject.com/en/4.2/howto/static-files/

STATIC_URL = 'static/'
STATICFILES_DIRS = [
    os.path.join(BASE_DIR, 'templates/static')
]

# Default primary key field type
# https://docs.djangoproject.com/en/4.2/ref/settings/#default-auto-field

DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'


# Tailwind CSS
TAILWIND_APP_NAME = 'app_tailwind'

INTERNAL_IPS = [
    "127.0.0.1",
]

# Verifica a plataforma atual
if platform.system() == "Windows":
    # Configura o caminho para o Windows
    NPM_BIN_PATH = r"C:\Program Files\nodejs\npm.cmd"
else:
    # Configura o caminho para Linux/Mac (assumindo que o Node.js e o NPM estão em /usr/local/bin/)
    NPM_BIN_PATH = "/usr/local/bin/npm"