requirements.txt
asgiref==3.7.2
Django==4.2.5
django-browser-reload==1.11.0
django-tailwind==3.6.0
gunicorn==21.2.0
packaging==23.2
python-decouple==3.8
sqlparse==0.4.4
pandas==2.2.2
numpy==1.22.4
